% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_reactions_by_intensity.R
\name{match_reactions_by_intensity}
\alias{match_reactions_by_intensity}
\title{Match reactions between two molecule datasets based on intensity filtering}
\usage{
match_reactions_by_intensity(file1, file2, reaction_delta_file, out_dir = ".")
}
\arguments{
\item{file1}{Path to the first molecular information CSV (e.g., inflow).}

\item{file2}{Path to the second molecular information CSV (e.g., outflow).}

\item{reaction_delta_file}{Path to the reaction delta CSV file.}

\item{out_dir}{Directory to save output files.}
}
\value{
Two CSV files saved in `out_dir`: `network_edge.csv` and `reaction_summary.csv`.
Match reactions between two molecule datasets based on intensity filtering

This function filters precursor and product molecules based on intensity changes,
and matches them with a list of possible reaction deltas.

Two CSV files saved in `out_dir`: `network_edge.csv` and `reaction_summary.csv`.
}
\description{
This function filters precursor and product molecules based on intensity changes,
and matches them with a list of possible reaction deltas.
}
