% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_reactions_by_mass_difference.R
\name{match_reactions_by_mass_difference}
\alias{match_reactions_by_mass_difference}
\title{Match reactions between two molecule datasets based on mass difference}
\usage{
match_reactions_by_mass_difference(
  file1,
  file2,
  reaction_delta_file,
  out_dir = ".",
  mass_tolerance = 0.005
)
}
\arguments{
\item{file1}{Path to the first molecular information CSV (e.g., inflow).}

\item{file2}{Path to the second molecular information CSV (e.g., outflow).}

\item{reaction_delta_file}{Path to the reaction delta CSV file (with mass differences).}

\item{out_dir}{Directory to save output files.}

\item{mass_tolerance}{Mass tolerance for matching (default: 0.005 Da).}
}
\value{
Two CSV files saved in `out_dir`: `network_edge.csv` and `reaction_summary.csv`.
}
\description{
This function filters precursor and product molecules based on intensity changes,
and matches them with a list of possible reaction mass differences.
}
