% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_mass_intensity}
\alias{merge_mass_intensity}
\title{Merge Mass and Intensity from Multiple CSV Files}
\usage{
merge_mass_intensity(
  dir_path,
  output_mass_intensity,
  output_mass_elements,
  output_meta_file = NULL
)
}
\arguments{
\item{dir_path}{The directory containing CSV files.}

\item{output_mass_intensity}{Path to save the merged Mass-Intensity file.}

\item{output_mass_elements}{Path to save the Mass-Elements file.}

\item{output_meta_file}{Path to save the auto-generated metadata file (optional).}
}
\value{
A peakData object with the merged data.
}
\description{
Merge Mass and Intensity from Multiple CSV Files
}
