% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_molform_intensity}
\alias{merge_molform_intensity}
\title{Merge Data Based on Molecular Formula (MolForm) with Filtering}
\usage{
merge_molform_intensity(
  dir_path,
  output_molform_intensity,
  output_molform_elements,
  output_filtered_samples_dir,
  output_meta_file = NULL
)
}
\arguments{
\item{dir_path}{The directory containing CSV files.}

\item{output_molform_intensity}{Path to save the merged MolForm-Intensity file.}

\item{output_molform_elements}{Path to save the MolForm-Elements file.}

\item{output_filtered_samples_dir}{Directory to save individual filtered sample files.}

\item{output_meta_file}{Path to save the auto-generated metadata file (optional).}
}
\value{
A peakData object with the merged data.
}
\description{
Merge Data Based on Molecular Formula (MolForm) with Filtering
}
