% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_molform_intensity}
\alias{merge_molform_intensity}
\title{Merge Molecular Formula and Intensity from Multiple CSV Files}
\usage{
merge_molform_intensity(
  dir_path,
  output_molform_intensity,
  output_molform_elements,
  output_meta_file = NULL
)
}
\arguments{
  \item{dir_path}{The directory containing CSV files.}

  \item{output_molform_intensity}{Path to save the merged MolForm-Intensity file.}

  \item{output_molform_elements}{Path to save the MolForm-Elements file.}

  \item{output_meta_file}{Path to save the auto-generated metadata file (optional).}
}
\value{
List containing merged data frames.
}
\description{
Merge molecular formulas and intensity data from multiple CSV files.
}
